backend/pyproject.toml
[project]
name = "my-agent-py-backend"
version = "0.1.0"
description = "Python backend for Couple-Agent (FastAPI, uv)"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "pgvector>=0.2.5",
    "redis>=5.0.0",
    "python-jose>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "alembic>=1.13.1",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=1.0.5",
    "langgraph-checkpoint-postgres>=2.0.0",
    "psycopg[binary,pool]>=3.1.0",
    "fastembed>=0.4.0",
    "tavily-python>=0.7.17",
    "langchain-google-genai>=4.0.0",
    "loguru>=0.7.3",
    "alibabacloud-oss-v2>=1.2.2",
    "langfuse>=2.0.0",
    # 知识库功能依赖
    "pypdf>=4.0.0",           # PDF 解析
    "python-docx>=1.1.0",     # Word 文档解析
    "docx2txt>=0.8",          # LangChain Docx2txtLoader 依赖
    "rank-bm25>=0.2.2",       # BM25 关键词检索
    "jieba>=0.42.1",          # 中文分词
    "langchain-community>=0.3.0",  # 社区 DocumentLoaders
    "langchain-text-splitters>=0.3.0",  # 文本分块器
    # Celery 消息队列
    "celery[redis]>=5.4.0",
    # 加密
    "cryptography>=43.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "black>=24.0.0",
    "mypy>=1.0.0",
]

[tool.uv]
constraint-dependencies = []


# Ruff 配置 - 代码检查与格式化
[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "W293",   # blank line contains whitespace (常见于 docstring)
    "C414",   # unnecessary list() call within sorted()
    "UP046", # Generic class uses old syntax (保持 Pydantic 兼容性)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

# Black 配置 - 代码格式化
[tool.black]
line-length = 100
target-version = ["py312"]

# Mypy 配置 - 类型检查
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true

# Pytest 配置
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
pythonpath = ["."]
